# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Craftive REST API",
        "description": "Dokumentasi interaktif REST API Craftive - Premium Handmade Goods Marketplace dengan integrasi Agentic AI Custom Planner. Dikembangkan sebagai Tugas Akhir Pemrograman API - UNESA 2026.",
        "contact": {
            "email": "dev@craftive.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server Lokal"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Registrasi Akun Baru",
                "description": "Mendaftarkan buyer atau seller baru. Setelah berhasil, token JWT langsung diberikan.",
                "operationId": "6703d7156c265a1e7b47e500d2a16304",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Siti Rahayu"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "siti@craftive.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "buyer",
                                        "enum": [
                                            "buyer",
                                            "seller"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil, token JWT diberikan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Login",
                "description": "Login dengan email dan password. Token JWT dikembalikan untuk dipakai di request selanjutnya.",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "siti@craftive.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil, token JWT diberikan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email atau password salah"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Profil Saya",
                "description": "Mendapatkan data user yang sedang login berdasarkan token JWT.",
                "operationId": "37252cdc2ae05f1e2e4fd541e987489d",
                "responses": {
                    "200": {
                        "description": "Data profil user berhasil diambil"
                    },
                    "401": {
                        "description": "Token tidak valid atau tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Logout",
                "description": "Keluar dan invalidasi token JWT.",
                "operationId": "0f6715df0d22b69c7e35bd32b4c1fc92",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Refresh Token",
                "description": "Memperbarui token JWT sebelum kadaluarsa.",
                "operationId": "d47f685292e0aa5af0c07fa89740bc36",
                "responses": {
                    "200": {
                        "description": "Token berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Lupa Password (Minta OTP)",
                "description": "Mengirim kode OTP 6 digit ke email untuk proses reset password.",
                "operationId": "95b6a0c8e3f351c53bf9d5da3ff40aef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "siti@craftive.id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP berhasil dikirim ke email"
                    },
                    "422": {
                        "description": "Email tidak terdaftar"
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Reset Password dengan OTP",
                "description": "Mengatur ulang password menggunakan kode OTP yang diterima via email.",
                "operationId": "8575b62977a43fb7fa1132f5bec5e6d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "siti@craftive.id"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diperbarui"
                    },
                    "422": {
                        "description": "OTP tidak valid atau kadaluarsa"
                    }
                }
            }
        },
        "/auth/profile": {
            "put": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Update Profil",
                "description": "Memperbarui nama, email, nomor telepon, dan alamat user yang sedang login.",
                "operationId": "53b368677faed6b1a65bd95dcba817b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Siti Rahayu"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "siti@craftive.id"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Batik No.5, Yogyakarta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dari endpoint Login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic Auth (email & password)",
                "scheme": "basic"
            }
        }
    },
...
}